+++ src/common/interfaces/modules/dynamic-module.interface.ts
@@ -1,0 +1,5 @@
+import { ModuleMetadata } from './module-metadata.interface';
+
+export interface DynamicModule extends ModuleMetadata {
+  module: any;
+}