+++ src/common/test/decorators/exception-filters.decorator.spec.ts
@@ -1,7 +1,7 @@
 import 'reflect-metadata';
 import { expect } from 'chai';
 import { EXCEPTION_FILTERS_METADATA } from '../../constants';
-import { UseFilters } from '../../utils/decorators/exception-filters.decorator';
+import { UseFilters } from '../../decorators/core/exception-filters.decorator';
 
 describe('@UseFilters', () => {
     const filters = [ 'exception', 'exception2' ];