+++ src/common/utils/bind-resolve-values.util.ts
@@ -1,6 +1,6 @@
 import { Constructor } from './merge-with-values.util';
 import { NestMiddleware } from '../interfaces/middlewares/nest-middleware.interface';
-import { Component } from './decorators/component.decorator';
+import { Component } from '../decorators/core/component.decorator';
 
 export const BindResolveMiddlewareValues = <T extends Constructor<NestMiddleware>>(data: Array<any>) => {
     return (Metatype: T): any => {