+++ src/core/test/injector/injector.spec.ts
@@ -2,7 +2,7 @@ import * as sinon from 'sinon';
 import { expect } from 'chai';
 import { InstanceWrapper } from '../../injector/container';
 import { Injector } from '../../injector/injector';
-import { Component } from '../../../common/utils/decorators/component.decorator';
+import { Component } from '../../../common/decorators/core/component.decorator';
 import { RuntimeException } from '../../errors/exceptions/runtime.exception';
 import { Module } from '../../injector/module';
 import { UnknownDependenciesException } from '../../errors/exceptions/unknown-dependencies.exception';